refactor: Page 레이어를 폴더링하여 적절하게 분리
