+++ src/index.js
@@ -1,7 +1,7 @@
-import "./style/index.css";
+import "./index.css";
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./pages/App";
+import App from "./pages/App/App";
 import { BrowserRouter } from "react-router-dom";
 
 ReactDOM.render(
