+++ src/pages/CardAddition/CardAddition.css
@@ -136,7 +136,7 @@
   display: inline-flex;
   width: 15rem;
   height: 10rem;
-  background-image: url("../asset/secure-code.png");
+  background-image: url("../../asset/secure-code.png");
   background-size: contain;
   background-repeat: no-repeat;
 }
