+++ src/pages/CompleteCardAddition/CompleteCardAddition.js
@@ -1,8 +1,9 @@
+import "./CompleteCardAddition.css";
 import React, { useState } from "react";
-import { CARD_SIZE } from "../constants";
+import { CARD_SIZE } from "../../constants";
 import PropTypes from "prop-types";
-import { Card, Button } from "../components";
-import { CARD_DESCRIPTION, FORMAT_CHAR, URL } from "../constants";
+import { Card, Button } from "../../components";
+import { CARD_DESCRIPTION, FORMAT_CHAR, URL } from "../../constants";
 import { Redirect } from "react-router";
 
 const CompleteCardAddition = (props) => {
